Realizamos la llamada ajax para las acciones de votar y quitar voto y su refactorización

+++ app/Repositories/VoteRepository.php
@@ -3,6 +3,7 @@
 namespace Curso\Repositories;
 
 
+use Curso\Entities\Ticket;
 use Curso\Entities\User;
 
 class VoteRepository
@@ -18,7 +19,11 @@ class VoteRepository
 
     public function unvote(User $user, Ticket $ticket)
     {
-        $user->voted()->detach($ticket);
+        if($user->hasVoted($ticket)) {
+            $user->voted()->detach($ticket);
+            return true;
+        }else
+            return false;
     }
 
 }
